+++ OwlCards/Cards/Curses/Envy.cs
@@ -1,0 +1,118 @@
+﻿using ModdingUtils.Extensions;
+using RarityBundle;
+using RarityLib.Utils;
+using UnityEngine;
+using System.Linq;
+using UnboundLib;
+using System.Collections;
+using OwlCards.Extensions;
+using OwlCards.Dependencies;
+
+namespace OwlCards.Cards.Curses
+{
+    internal class Envy : AOwlCard
+    {
+        public override void SetupCard_child(CardInfo cardInfo, Gun gun, ApplyCardStats cardStats, CharacterStatModifiers statModifiers, Block block)
+        {
+			conditions[GetTitle()] = (float soul) => { return soul < 1.5; };//Mismatch is intended, to put player in negative
+			if (!cardInfo.categories.Contains(CurseHandler.CurseCategory))
+				cardInfo.categories = cardInfo.categories.Append(CurseHandler.CurseCategory).ToArray();
+			if (!cardInfo.categories.Contains(OwlCardCategory.soulCondition))
+				cardInfo.categories = cardInfo.categories.Append(OwlCardCategory.soulCondition).ToArray();
+            //Edits values on card itself, which are then applied to the player in `ApplyCardStats`
+        }
+        public override void OnAddCard(Player player, Gun gun, GunAmmo gunAmmo, CharacterData data, HealthHandler health, Gravity gravity, Block block, CharacterStatModifiers characterStats)
+        {
+			if (CurseHandler.bCurseAvailable)
+			{
+				OwlCards.instance.ExecuteAfterFrames(20, () =>
+				{
+					foreach (int otherPlayerID in Utils.GetOpponentsPlayersIDs(player.playerID))
+					{
+						float otherPlayerSoul = OwlCardsData.GetData(otherPlayerID).Soul;
+						int cursesToGive = 1;
+						if (otherPlayerSoul > 2)
+							cursesToGive++;
+						if (otherPlayerSoul > 3.5)
+							cursesToGive++;
+						OwlCards.instance.StartCoroutine(GiveCurses(Utils.GetPlayerWithID(otherPlayerID), cursesToGive));
+					}
+				});
+			}
+            //Edits values on player when card is selected
+        }
+
+		private IEnumerator GiveCurses(Player player, int amount)
+		{
+			for (int i = 0; i < amount; i++)
+			{
+				CurseHandler.CursePlayer(player, (curse) => { ModdingUtils.Utils.CardBarUtils.instance.ShowAtEndOfPhase(player, curse); });
+				for (int j = 0; j < 20; j++)
+					yield return null;
+			}
+			yield break;
+		}
+
+        public override void OnRemoveCard(Player player, Gun gun, GunAmmo gunAmmo, CharacterData data, HealthHandler health, Gravity gravity, Block block, CharacterStatModifiers characterStats)
+        {
+            //Run when the card is removed from the player
+        }
+
+        protected override string GetTitle()
+        {
+            return "Envy";
+        }
+        protected override string GetDescription()
+        {
+            return "All foes get curses the more soul they have";
+        }
+        protected override CardInfoStat[] GetStats()
+        {
+            return new CardInfoStat[]
+            {
+                new CardInfoStat()
+                {
+                    positive = true,
+                    stat = "Curse to Foes",
+                    amount = "+1",
+                    simepleAmount = CardInfoStat.SimpleAmount.notAssigned
+                },
+                new CardInfoStat()
+                {
+                    positive = true,
+                    stat = "Curse to Foes with 2 Soul",
+                    amount = "+1",
+                    simepleAmount = CardInfoStat.SimpleAmount.notAssigned
+                },
+                new CardInfoStat()
+                {
+                    positive = true,
+                    stat = "Curse to Foes with 3.5 Soul",
+                    amount = "+1",
+                    simepleAmount = CardInfoStat.SimpleAmount.notAssigned
+                },
+                new CardInfoStat()
+                {
+                    positive = false,
+                    stat = "Soul",
+                    amount = "-2",
+                    simepleAmount = CardInfoStat.SimpleAmount.notAssigned
+                }
+            };
+        }
+
+        protected override GameObject GetCardArt()
+        {
+            return GetCardArt("C_Envy");
+        }
+        protected override CardInfo.Rarity GetRarity()
+        {
+            return Rarities.Uncommon;
+        }
+
+        protected override CardThemeColor.CardThemeColorType GetTheme()
+        {
+            return CardThemeColor.CardThemeColorType.EvilPurple;
+        }
+    }
+}